refactor(common-utils): move dashboard validators into their own module
Review feedback from pulpdrew on #2201:

- `validateDashboardContainersStructure` and
  `validateDashboardTileContainerRefs` (and the two helper types
  `ContainerForValidation` / `TileForValidation`) move out of
  `@hyperdx/common-utils/dist/types` into a new
  `@hyperdx/common-utils/dist/dashboardValidation` module. The types
  file is back to just types and type guards, matching the rest of the
  codebase.
- Drop `validateDashboardContainersConsistency`. It was only referenced
  by its own unit test; production code in the v2 dashboards router
  calls the two underlying helpers in sequence directly (and has to,
  because the structure pass runs at the schema level while the
  tile-ref pass runs at the handler level so PUT can fall back to the
  existing dashboard's containers).
- Test file now exercises the two helpers in the same order production
  does, with a `runHelper` that wires them into one `z.superRefine`.
  All 19 existing assertions hold unchanged.

No behaviour change for external API callers.
